reorganized component organization and centered blog posts
