+++ frontend/src/components/Blog/Post/Post.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function Post() {
+    return (
+        <div className='m-4 w-64 h-64 bg-purple-500'>
+            <p>This is a post!</p>
+        </div>
+    );
+}
+
+export default Post;
